requirements.txt
Flask
gunicorn
sklearn
numpy
pandas
PyYAML
-e .